fix: move MCP server creation logic to mcp-server.ts to resolve circular dependency issue
